app/ai_css_generator/pyproject.toml
[tool.poetry]
name = "prompt-guard"
version = "0.1.0"
description = ""
authors = ["Твой Ник <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = { extras = ["standard"], version = "^0.29" }
pydantic = "^2.7"
langdetect = "^1.0.9"
tenacity = "^8.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
